requirements.txt
python-telegram==0.12.0
